Added support for user input values for PI controller and cooking temp

+++ ASE6030/SequenceParameters.cs
@@ -12,7 +12,7 @@ namespace ASE6030
 
         // Cooking
         public int cookingTime;
-        public int cookingTemperature;
+        public double cookingTemperature;
         public int cookingPressure;
 
         // PI Controller
